schemars::transform

Trait Transform

pub trait Transform {
    // Required method
    fn transform(&mut self, schema: &mut Schema);
}

Trait used to modify a constructed schema and optionally its subschemas.

See the module documentation for more details on implementing this trait.

Required Methods

fn transform(&mut self, schema: &mut Schema)

Applies the transform to the given Schema.

When overriding this method, you may want to call the transform_subschemas function to also transform any subschemas.