Type Definition schemars::MapEntry

pub type MapEntry<'a, K, V> = Entry<'a, K, V>;

A view into a single entry in a map, which may either be vacant or occupied. This is constructed from the entry method on BTreeMap or IndexMap, depending on whether the preserve_order feature flag is set.