schemars

Module gen

JSON Schema generator and settings.

This module is useful if you want more control over how the schema generated than the schema_for! macro gives you. There are two main types in this module:

SchemaSettings, which defines what JSON Schema features should be used when generating schemas (for example, how Options should be represented).

SchemaGenerator, which manages the generation of a schema document.

Structs

SchemaGenerator

The main type used to generate JSON Schemas.

SchemaSettings

Settings to customize how Schemas are generated.

Traits

GenVisitor

A Visitor which implements additional traits required to be included in a SchemaSettings.