Struct schemars::gen::SchemaGenerator

pub struct SchemaGenerator { /* private fields */ }

The main type used to generate JSON Schemas.

Example

use schemars::{JsonSchema, gen::SchemaGenerator};

#[derive(JsonSchema)]
struct MyStruct {
    foo: i32,
}

let gen = SchemaGenerator::default();
let schema = gen.into_root_schema_for::<MyStruct>();

Implementations

impl SchemaGenerator

pub fn new(settings: SchemaSettings) -> SchemaGenerator

Creates a new SchemaGenerator using the given settings.

pub fn settings(&self) -> &SchemaSettings

Borrows the SchemaSettings being used by this SchemaGenerator.

Example

use schemars::gen::SchemaGenerator;

let gen = SchemaGenerator::default();
let settings = gen.settings();

assert_eq!(settings.option_add_null_type, true);

pub fn make_extensible(&self, _schema: &mut SchemaObject)

๐Deprecated: This method no longer has any effect.

pub fn schema_for_any(&self) -> Schema

๐Deprecated: Use Schema::Bool(true) instead

pub fn schema_for_none(&self) -> Schema

๐Deprecated: Use Schema::Bool(false) instead

pub fn subschema_for<T: ?Sized + JsonSchema>(&mut self) -> Schema

Generates a JSON Schema for the type T, and returns either the schema itself or a $ref schema referencing Tโs schema.

If T is referenceable, this will add Tโs schema to this generatorโs definitions, and return a $ref schema referencing that schema. Otherwise, this method behaves identically to JsonSchema::json_schema.

If Tโs schema depends on any referenceable schemas, then this method will add them to the SchemaGeneratorโs schema definitions.

pub fn definitions(&self) -> &Map<String, Schema>

Borrows the collection of all referenceable schemas that have been generated.

The keys of the returned Map are the schema names, and the values are the schemas themselves.

pub fn definitions_mut(&mut self) -> &mut Map<String, Schema>

Mutably borrows the collection of all referenceable schemas that have been generated.

The keys of the returned Map are the schema names, and the values are the schemas themselves.

pub fn take_definitions(&mut self) -> Map<String, Schema>

Returns the collection of all referenceable schemas that have been generated, leaving an empty map in its place.

The keys of the returned Map are the schema names, and the values are the schemas themselves.

pub fn visitors_mut(&mut self) -> impl Iterator<Item = &mut dyn GenVisitor>

Returns an iterator over the visitors being used by this SchemaGenerator.

pub fn root_schema_for<T: ?Sized + JsonSchema>(&mut self) -> RootSchema

Generates a root JSON Schema for the type T.

If Tโs schema depends on any referenceable schemas, then this method will add them to the SchemaGeneratorโs schema definitions and include them in the returned SchemaObjectโs definitions

pub fn into_root_schema_for<T: ?Sized + JsonSchema>(self) -> RootSchema

Consumes self and generates a root JSON Schema for the type T.

If Tโs schema depends on any referenceable schemas, then this method will include them in the returned SchemaObjectโs definitions

pub fn root_schema_for_value<T: ?Sized + Serialize>( &mut self, value: &T ) -> Result<RootSchema, Error>

Generates a root JSON Schema for the given example value.

If the value implements JsonSchema, then prefer using the root_schema_for() function which will generally produce a more precise schema, particularly when the value contains any enums.

pub fn into_root_schema_for_value<T: ?Sized + Serialize>( self, value: &T ) -> Result<RootSchema, Error>

Consumes self and generates a root JSON Schema for the given example value.

If the value implements JsonSchema, then prefer using the into_root_schema_for()! function which will generally produce a more precise schema, particularly when the value contains any enums.

pub fn dereference<'a>(&'a self, schema: &Schema) -> Option<&'a Schema>

Attemps to find the schema that the given schema is referencing.

If the given schema has a $ref property which refers to another schema in selfโs schema definitions, the referenced schema will be returned. Otherwise, returns None.

Example

use schemars::{JsonSchema, gen::SchemaGenerator};

#[derive(JsonSchema)]
struct MyStruct {
    foo: i32,
}

let mut gen = SchemaGenerator::default();
let ref_schema = gen.subschema_for::<MyStruct>();

assert!(ref_schema.is_ref());

let dereferenced = gen.dereference(&ref_schema);

assert!(dereferenced.is_some());
assert!(!dereferenced.unwrap().is_ref());
assert_eq!(dereferenced, gen.definitions().get("MyStruct"));