Struct schemars::gen::SchemaSettings

pub struct SchemaSettings {
    pub option_nullable: bool,
    pub option_add_null_type: bool,
    pub definitions_path: String,
    pub meta_schema: Option<String>,
    pub visitors: Vec<Box<dyn GenVisitor>>,
    pub inline_subschemas: bool,
    // some fields omitted
}

Settings to customize how Schemas are generated.

The default settings currently conform to JSON Schema Draft 7, but this is liable to change in a future version of Schemars if support for other JSON Schema versions is added. If you require your generated schemas to conform to draft 7, consider using the draft07 method.

Fields
option_nullable: bool

If true, schemas for Option<T> will include a nullable property.

This is not part of the JSON Schema spec, but is used in Swagger/OpenAPI schemas.

Defaults to false.

option_add_null_type: bool

If true, schemas for Option<T> will have null added to their type.

Defaults to true.

definitions_path: String

A JSON pointer to the expected location of referenceable subschemas within the resulting root schema.

Defaults to "#/definitions/".

meta_schema: Option<String>

The URI of the meta-schema describing the structure of the generated schemas.

Defaults to "http://json-schema.org/draft-07/schema#".

visitors: Vec<Box<dyn GenVisitor>>

A list of visitors that get applied to all generated root schemas.

inline_subschemas: bool

Inline all subschemas instead of using references.

Some references may still be generated in schemas for recursive types.

Defaults to false.

Implementations

impl SchemaSettings

pub fn draft07() -> SchemaSettings

Creates SchemaSettings that conform to JSON Schema Draft 7.

pub fn draft2019_09() -> SchemaSettings

Creates SchemaSettings that conform to JSON Schema 2019-09.

pub fn openapi3() -> SchemaSettings

Creates SchemaSettings that conform to OpenAPI 3.0.

pub fn with(self, configure_fn: impl FnOnce(&mut Self)) -> Self

Modifies the SchemaSettings by calling the given function.

Example

use schemars::gen::{SchemaGenerator, SchemaSettings};

let settings = SchemaSettings::default().with(|s| {
    s.option_nullable = true;
    s.option_add_null_type = false;
});
let gen = settings.into_generator();

pub fn with_visitor(
self,
visitor: impl Visitor + Debug + Clone + 'static
) -> Self

Appends the given visitor to the list of visitors for these SchemaSettings.

pub fn into_generator(self) -> SchemaGenerator

Creates a new SchemaGenerator using these settings.