Trait schemars::gen::GenVisitor

pub trait GenVisitor: Visitor + Debug + DynClone + Any {
    fn as_any(&self) -> &dyn Any;
}

A Visitor which implements additional traits required to be included in a SchemaSettings.

You will rarely need to use this trait directly as it is automatically implemented for any type which implements all of:

Visitor

std::fmt::Debug

std::any::Any (implemented for all 'static types)

std::clone::Clone

Example

use schemars::visit::Visitor;
use schemars::gen::GenVisitor;

#[derive(Debug, Clone)]
struct MyVisitor;

impl Visitor for MyVisitor { }

let v: &dyn GenVisitor = &MyVisitor;
assert!(v.as_any().is::<MyVisitor>());

Required Methods

fn as_any(&self) -> &dyn Any

Upcasts this visitor into an Any, which can be used to inspect and manipulate it as its concrete type.