Macro schemars::schema_for_value

macro_rules! schema_for_value {
    ($value:expr) => { ... };
}

Generates a RootSchema for the given example value using default settings.

The value must implement Serialize. If the value also implements JsonSchema, then prefer using the schema_for! macro which will generally produce a more precise schema, particularly when the value contains any enums.

If the Serialize implementation of the value decides to fail, this macro will panic. For a non-panicking alternative, create a SchemaGenerator and use its into_root_schema_for_value method.

Example

use schemars::schema_for_value;

#[derive(serde::Serialize)]
struct MyStruct {
    foo: i32,
}

let my_schema = schema_for_value!(MyStruct { foo: 123 });