Module schemars::schema

JSON Schema types.

Structs

ArrayValidation

Properties of a SchemaObject which define validation assertions for arrays.

Metadata

Properties which annotate a SchemaObject which typically have no effect when an object is being validated against the schema.

NumberValidation

Properties of a SchemaObject which define validation assertions for numbers.

ObjectValidation

Properties of a SchemaObject which define validation assertions for objects.

RootSchema

The root object of a JSON Schema document.

SchemaObject

A JSON Schema object.

StringValidation

Properties of a SchemaObject which define validation assertions for strings.

SubschemaValidation

Properties of a SchemaObject which define validation assertions in terms of other schemas.

Enums

InstanceType

The possible types of values in JSON Schema documents.

Schema

A JSON Schema.

SingleOrVec

A type which can be serialized as a single item, or multiple items.