Trait schemars::visit::Visitor

pub trait Visitor {
    fn visit_root_schema(&mut self, root: &mut RootSchema) { ... }

    fn visit_schema(&mut self, schema: &mut Schema) { ... }

    fn visit_schema_object(&mut self, schema: &mut SchemaObject) { ... }
}

Trait used to recursively modify a constructed schema and its subschemas.

Provided methods

fn visit_root_schema(&mut self, root: &mut RootSchema)

Override this method to modify a RootSchema and (optionally) its subschemas.

When overriding this method, you will usually want to call the visit_root_schema function to visit subschemas.

fn visit_schema(&mut self, schema: &mut Schema)

Override this method to modify a Schema and (optionally) its subschemas.

When overriding this method, you will usually want to call the visit_schema function to visit subschemas.

fn visit_schema_object(&mut self, schema: &mut SchemaObject)

Override this method to modify a SchemaObject and (optionally) its subschemas.

When overriding this method, you will usually want to call the visit_schema_object function to visit subschemas.

Loading content...